+++ app/src/main/java/com/fibelatti/photowidget/platform/IntentKtx.kt
@@ -5,6 +5,8 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Build
 import android.provider.Settings
+import androidx.core.content.FileProvider
+import java.io.File
 
 fun Intent.setIdentifierCompat(value: String?): Intent = apply {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
@@ -26,4 +28,35 @@ fun appSettingsIntent(context: Context): Intent {
 
 fun batteryUsageSettingsIntent(): Intent {
     return Intent(Settings.ACTION_IGNORE_BATTERY_OPTIMIZATION_SETTINGS)
+}
+
+fun sharePhotoChooserIntent(
+    context: Context,
+    originalPhotoPath: String?,
+    externalUri: Uri?,
+): Intent? {
+    val uri: Uri
+    when {
+        originalPhotoPath != null -> {
+            uri = FileProvider.getUriForFile(
+                /* context = */ context,
+                /* authority = */ "${context.packageName}.fileprovider",
+                /* file = */ File(originalPhotoPath),
+            )
+        }
+
+        externalUri != null -> {
+            uri = externalUri
+        }
+
+        else -> {
+            return null
+        }
+    }
+
+    val shareIntent: Intent = Intent(Intent.ACTION_SEND)
+        .putExtra(Intent.EXTRA_STREAM, uri)
+        .setType("image/*")
+
+    return Intent.createChooser(shareIntent, null)
 }